+++ internal/linux/linux.go
@@ -74,9 +74,9 @@ func Sendmsg(fd int, p, oob []byte, to unix.Sockaddr, flags int) error {
 }
 
 // SetMempolicy wraps set_mempolicy.
-func SetMempolicy(mode int, mask *unix.CPUSet) error {
+func SetMempolicy(mode int, mask unix.CPUSetDynamic) error {
 	err := retryOnEINTR(func() error {
-		return unix.SetMemPolicy(mode, mask)
+		return unix.SetMemPolicyDynamic(mode, mask)
 	})
 	return os.NewSyscallError("set_mempolicy", err)
 }