+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### libcontainer API ###
+- `configs.ToCPUSet` now returns a `unix.CPUSetDynamic` instead of a
+  `*unix.CPUSet`, and the `Initial`/`Final` fields of `configs.CPUAffinity` and
+  the `Nodes` field of `configs.LinuxMemoryPolicy` have changed type
+  accordingly. This lifts the previous 1024 CPUs/nodes limit. (#5343)
+
 ### Fixed ###
 - The poststart hooks are now executed after starting the user-specified
   process, fixing a runtime-spec conformance issue. (#4347, #5186)
@@ -17,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed ###
 - runc now depends on [libpathrs v0.2.5] or later, and attempting to build with
   older versions will cause compilation errors. (#5291)
+- The `cpuAffinity` and NUMA `memoryPolicy` settings are no longer limited to
+  1024 CPUs/nodes, as runc now uses a dynamically-sized CPU mask. (#5343)
 
 [libpathrs v0.2.5]: https://github.com/cyphar/libpathrs/releases/tag/v0.2.5